LayoutTests/platform/mac/accessibility/legend-expected.txt

Choose a shipping method:
Overnight
End of test
This tests that a fieldset's title ui element is the legend.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


AXRole: AXWebArea AXValue: 
    AXRole: AXGroup AXValue: 
        AXRole: AXGroup AXValue:      << fieldset's titleUIElement
            AXRole: AXStaticText AXValue: Choose a shipping method:
        AXRole: AXGroup AXValue: 
            AXRole: AXRadioButton AXValue: 0
            AXRole: AXStaticText AXValue: Overnight
    AXRole: AXGroup AXValue: 
        AXRole: AXStaticText AXValue: End of test
PASS titleUIElement != null is true
PASS titleUIElementText.stringValue is 'AXValue: Choose a shipping method:'
PASS successfullyParsed is true

TEST COMPLETE
